+++ src/rgb_xyz.h
@@ -31,7 +31,7 @@ extern boost::shared_ptr<ARGBFrame> xyz_to_rgb (
 	);
 
 extern boost::shared_ptr<XYZFrame> rgb_to_xyz (
-	boost::shared_ptr<const Image>, boost::shared_ptr<const LUT>, boost::shared_ptr<const LUT>, float const colour_matrix[3][3]
+	boost::shared_ptr<const Image>, boost::shared_ptr<const LUT>, boost::shared_ptr<const LUT>, double const colour_matrix[3][3]
 	);
 	
 }
